Improve menu buttons and fix key commands
