+++ client/features/editor/components/EditorControls/HistoryControls/HistoryControls.tsx
@@ -1,19 +1,13 @@
-import {
-  ArrowCircleLeftIcon,
-  ArrowLeftIcon,
-  ReplyIcon,
-} from '@heroicons/react/outline';
+import { ReplyIcon } from '@heroicons/react/outline';
 import React from 'react';
 import { useRecoilValue } from 'recoil';
-import RedoIcon from '../../../../components/ui/Icons/RedoIcon';
-import UndoIcon from '../../../../components/ui/Icons/UndoIcon';
-import Tooltip from '../../../../components/ui/Tooltip/Tooltip';
-import useHistoryDispatcher from '../../state/dispatchers/history';
+import Tooltip from '../../../../../components/ui/Tooltip/Tooltip';
+import useHistoryDispatcher from '../../../state/dispatchers/history';
 import {
   canRedoSelector,
   canUndoSelector,
-} from '../../state/selectors/history';
-import ClearButton from '../ui/ClearButton';
+} from '../../../state/selectors/history';
+import ClearButton from '../../ui/ClearButton';
 
 function HistoryControls() {
   const canUndo = useRecoilValue(canUndoSelector);
