+++ client/features/editor/components/ui/SideMenuButton.tsx
@@ -30,13 +30,16 @@ function SideMenuButton({
       disabled={disabled}
       onClick={disabled ? undefined : onClick}
       className={classNames(
-        'relative h-20 w-20 p-2 text-xs flex flex-col items-center justify-center rounded-md border focus:outline-none transition duration-150',
+        'relative h-20 w-20 p-2 text-xs flex flex-col items-center justify-center rounded-md shadow-sm hover:shadow-none border focus:outline-none transition duration-150',
         selected
           ? 'bg-blue-50 text-blue-800 border-blue-300'
-          : 'bg-white text-gray-700 hover:border-gray-400 hover:text-gray-900 focus:border-gray-400 focus:text-gray-900',
+          : 'bg-white text-gray-700',
         selected &&
           !disabled &&
           'hover:border-blue-400 hover:text-blue-900 focus:border-blue-400 focus:text-blue-900',
+        !selected &&
+          !disabled &&
+          'bg-gray-50 hover:border-gray-400 hover:text-gray-900 focus:border-gray-400 focus:text-gray-900',
         disabled && 'cursor-default',
         className
       )}
